Run consistency tests in float64
Switch the consistency-test inference precision from float32 to float64
and regenerate the darwin_arm64 reference predictions accordingly.
float64 reduces floating-point divergence across hardware/BLAS backends,
making the reference predictions more stable across machines.

+++ tests/test_consistency.py
@@ -119,7 +119,9 @@ DEFAULT_CONFIG = {
     "n_estimators": 2,
     "random_state": 42,
     "device": "cpu",
-    "inference_precision": torch.float32,
+    # float64 minimises floating-point divergence across hardware/BLAS backends,
+    # keeping predictions comparable to the stored references.
+    "inference_precision": torch.float64,
 }
 
 TEST_CASES = {
